Implement craneOperator9001 behavior
Fix test for craneOperator9001 as well as I had (again) added the crates
in the wrong order.

This modification was a bit easier for me to get done, but the
crappy/verbose syntax, along with the `-1` strewn everywhere made it
harder for me to notice that my "pop" was only pulling one, rather than
many off of the stack. Again, this is probably something that would have
been more easily manageable with as a linked list and that's probably
the first thing I'm coming back for when I wrap this year up.